refactor hooks for new api
